third_party/WebKit/LayoutTests/media/encrypted-media/encrypted-media-unprefixed-after-prefixed-expected.txt

Call unprefixed EME API after prefixed EME API.

EVENT(loadstart)
This is a testharness.js-based test.
PASS Call unprefixed EME API after prefixed EME API. 
Harness: the test ran to completion.
